requirements.txt
requests>=2.28.0
typing-extensions>=4.4.0
python-dotenv>=1.0.0
pytest==8.1.1
pytest-mock==3.12.0
